Separate src/detections/rule/mod.rs into a rulenode module
mod.rs had grown to ~1200 lines that mixed module declarations with the
core rule types (RuleNode, DetectionNode, AggResult, CorrelationType),
the create_rule/get_detection_keys helpers, and a ~690-line test module,
making the module layout hard to read (issue #1045).

Moved all of those items into a new src/detections/rule/rulenode.rs,
leaving mod.rs as pure module wiring plus `pub use rulenode::*;`. The
re-export keeps every existing `crate::detections::rule::{RuleNode, ...}`
path working unchanged. The few members that sibling modules reach
(RuleNode::new_with_detection, RuleNode::countdata, DetectionNode::new,
the shared test helpers) are widened from private to pub(super)/pub(crate)
to preserve the exact access they had as descendants of `rule` — a
compile-time-only change.

Pure code movement, no behavior change: detection output is byte-identical
on the sample-evtx corpus, and cargo fmt/clippy -D warnings/test all pass
(512 tests).
